+++ src/plugins/WebOfTrust/util/BackgroundJob.java
@@ -27,6 +27,7 @@ public interface BackgroundJob {
      * This method does not wait for the job to terminate, it only triggers its termination and
      * returns immediately. Use this method in conjunction with {@link #waitForTermination(long)}
      * when blocking behaviour is required.
+     * Implementations must ensure that this method is safe to invoke from any thread at any moment.
      * @see #isTerminated()
      * @see #waitForTermination(long)
      */
@@ -37,6 +38,7 @@ public interface BackgroundJob {
      * of {@link #triggerExecution()} have no effect. Implementations must ensure that a terminated
      * background job cannot be restarted, i.e. once this method returns {@code true}, it will
      * always return {@code true}.
+     * Implementations must ensure that this method is safe to invoke from any thread at any moment.
      * @return {@code true} if this background job is terminated
      * @see #terminate()
      */
@@ -47,6 +49,7 @@ public interface BackgroundJob {
      * the thread is interrupted.
      * This method does not terminate the background job, it only waits for it to terminate. Use
      * this method in conjunction with {@link #terminate()} when terminating behaviour is required.
+     * Implementations must ensure that this method is safe to invoke from any thread at any moment.
      * @param timeout the maximum time to wait in milliseconds
      * @throws InterruptedException When the calling thread is interrupted.
      * @see #terminate()
